# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PTSite API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/v1/nights/{night}/attendance/{player}": {
            "put": {
                "operationId": "attendance.update",
                "summary": "Set a player's answer: \"all_in\" or \"fold\". Players answer for themselves; results keepers and admins for\nanyone. Repeating the same answer changes nothing",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "night",
                        "in": "path",
                        "required": true,
                        "description": "The night ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "player",
                        "in": "path",
                        "required": true,
                        "description": "The player ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnswerAttendanceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AttendanceResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AttendanceResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "attendance.destroy",
                "summary": "Remove a player's answer (\"N\u00e3o confirmado\")",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "night",
                        "in": "path",
                        "required": true,
                        "description": "The night ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "player",
                        "in": "path",
                        "required": true,
                        "description": "The player ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/holiday-calendar/{year}": {
            "get": {
                "operationId": "holiday.calendar",
                "summary": "The holidays of one year, in date order, including the ones cancelled for that year and the extras",
                "tags": [
                    "Holiday"
                ],
                "parameters": [
                    {
                        "name": "year",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `CalendarHolidayResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CalendarHolidayResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/v1/holiday-exceptions": {
            "post": {
                "operationId": "holiday.storeException",
                "summary": "Cancel a table holiday for one year, or add an extra holiday that year",
                "tags": [
                    "Holiday"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveHolidayExceptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "`HolidayExceptionResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/HolidayExceptionResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/v1/holiday-exceptions/{exception}": {
            "delete": {
                "operationId": "holiday.destroyException",
                "summary": "Undo a one-year change",
                "tags": [
                    "Holiday"
                ],
                "parameters": [
                    {
                        "name": "exception",
                        "in": "path",
                        "required": true,
                        "description": "The exception ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/seasons/{season}/calendar": {
            "get": {
                "operationId": "season.calendar",
                "summary": "The season calendar: its nights (with the winner of finished ones, the ALL IN count and the user's answer)\nand the regular nights left out because of a holiday, an emenda or Carnival, in date order",
                "tags": [
                    "Season"
                ],
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "description": "The season ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `CalendarEntryResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CalendarEntryResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "AnswerAttendanceRequest": {
                "type": "object",
                "properties": {
                    "answer": {
                        "description": "\"all_in\" (coming) or \"fold\" (not coming).",
                        "$ref": "#/components/schemas/AttendanceAnswer"
                    }
                },
                "required": [
                    "answer"
                ],
                "title": "AnswerAttendanceRequest"
            },
            "AttendanceAnswer": {
                "type": "string",
                "description": "A player's answer for a night: coming (\"ALL IN\") or not (\"FOLD\").\n",
                "enum": [
                    "all_in",
                    "fold"
                ],
                "title": "AttendanceAnswer"
            },
            "AttendanceResource": {
                "type": "object",
                "properties": {
                    "player": {
                        "$ref": "#/components/schemas/PlayerResource"
                    },
                    "answer": {
                        "type": "string",
                        "enum": [
                            "all_in",
                            "fold"
                        ]
                    },
                    "answered_at": {
                        "type": "string"
                    },
                    "answered_by": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "description": "Who set the answer, when it was not the player: a results keeper or admin.",
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "id",
                            "name"
                        ]
                    }
                },
                "required": [
                    "player",
                    "answer",
                    "answered_at",
                    "answered_by"
                ],
                "title": "AttendanceResource"
            },
...
            "CalendarEntryResource": {
                "type": "object",
                "properties": {
                    "kind": {
                        "type": "string",
                        "enum": [
                            "night",
                            "no_night"
                        ]
                    },
                    "starts_at": {
                        "type": "string"
                    },
                    "night": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "description": "Set for nights.",
                        "properties": {
                            "id": {
                                "type": [
                                    "integer",
                                    "null"
                                ]
                            },
                            "status": {
                                "type": "string",
                                "enum": [
                                    "scheduled",
                                    "open",
                                    "finished"
                                ]
                            },
                            "place": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "winner": {
                                "type": [
                                    "string",
                                    "null"
                                ],
                                "description": "Nickname of the 1st place, for finished nights."
                            },
                            "pot": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "all_in_count": {
                                "type": "integer",
                                "description": "How many players answered ALL IN."
                            },
                            "my_answer": {
                                "type": [
                                    "string",
                                    "null"
                                ],
                                "enum": [
                                    "all_in",
                                    "fold",
                                    null
                                ]
                            }
                        },
                        "required": [
                            "id",
                            "status",
                            "place",
                            "winner",
                            "pot",
                            "all_in_count",
                            "my_answer"
                        ]
                    },
                    "skip_reason": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "description": "Set for regular nights left out: holiday, bridge (the day before is a holiday) or carnival.",
                        "properties": {
                            "kind": {
                                "anyOf": [
                                    {
                                        "$ref": "#/components/schemas/SkipKind"
                                    },
                                    {
                                        "type": "null"
                                    }
                                ]
                            },
                            "holiday": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        },
                        "required": [
                            "kind",
                            "holiday"
                        ]
                    }
                },
                "required": [
                    "kind",
                    "starts_at",
                    "night",
                    "skip_reason"
                ],
                "title": "CalendarEntryResource"
            },
            "CalendarHolidayResource": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "scope": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "holiday_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "cancelled": {
                        "type": "boolean"
                    },
                    "exception_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    }
                },
                "required": [
                    "date",
                    "name",
                    "scope",
                    "holiday_id",
                    "cancelled",
                    "exception_id"
                ],
                "title": "CalendarHolidayResource"
            },
...
            "HolidayExceptionResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "holiday_id": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "description": "Set when a table holiday is cancelled for the year."
                    },
                    "date": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Set for an extra holiday that year (Y-m-d)."
                    },
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "year",
                    "holiday_id",
                    "date",
                    "name"
                ],
                "title": "HolidayExceptionResource"
            },
...
            "PlayerResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nickname": {
                        "type": "string"
                    },
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "status": {
                        "type": "string"
                    },
                    "archived": {
                        "type": "boolean"
                    },
                    "thumbnail_version": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Null when the player has no thumbnail. The image is at /players/{id}/thumbnail; add ?v=<this value>\nto the URL, because the answer is cached for a year."
                    },
                    "photo_version": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The same for the larger photo, at /players/{id}/photo."
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Contact details only for admins and the player themself."
                    },
                    "birth_date": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "login": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "description": "Site access, only for admins: null when the player has no login.",
                        "properties": {
                            "username": {
                                "type": "string"
                            },
                            "role": {
                                "type": "string",
                                "enum": [
                                    "player",
                                    "results_keeper",
                                    "admin"
                                ]
                            },
                            "last_login_at": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        },
                        "required": [
                            "username",
                            "role",
                            "last_login_at"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "nickname",
                    "name",
                    "status",
                    "archived",
                    "thumbnail_version",
                    "photo_version",
                    "updated_at"
                ],
                "title": "PlayerResource"
            },
...
            "SaveHolidayExceptionRequest": {
                "type": "object",
                "description": "Either holiday_id (cancel that holiday for the year) or date and name (an extra holiday that year).",
                "properties": {
                    "year": {
                        "type": "integer",
                        "minimum": 1900,
                        "maximum": 2200
                    },
                    "holiday_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "date": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date"
                    },
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 100
                    }
                },
                "required": [
                    "year"
                ],
                "title": "SaveHolidayExceptionRequest"
            },
...
            "SkipKind": {
                "type": "string",
                "description": "Why the season planner leaves out a regular night.\n| |\n|---|\n| `holiday` <br/> The night itself is a holiday. |\n| `bridge` <br/> The day before is a holiday (\"emenda\"): a long weekend. |\n| `carnival` <br/> The night falls in the Carnival weekend. |",
                "enum": [
                    "holiday",
                    "bridge",
                    "carnival"
                ],
                "title": "SkipKind"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}